fixed a bug in filter(..Neuromodulator)
this bug caused data with more than one neuromodulator to
slip through the filter. this was because adding two logical
true values gives you true, not 2. fixed by casting to double

+++ +embedding/@DataStore/filter.m
@@ -78,7 +78,7 @@ case sourcedata.DataFilter.Neuromodulator
 	% at the same time
 	N_mod = zeros(length(data.mask),1);
 	for modulator = List(modulators)
-		N_mod = N_mod + data.(modulator)>0;
+		N_mod = N_mod + double(data.(modulator)>0);
 	end
 	if any(N_mod>1)
 		data = data.purge(N_mod>1);
